fixes for stting admin username and password, added support for v2 and above
fixes for unit tests

fix to rubocop

+++ manifests/init.pp
@@ -59,12 +59,20 @@
 #   The HTTPS URL for the UCP controller, used by nodes to join the cluster.
 #   Required for nodes.
 #
+# [*ucp_manager*]
+#   The ip address of the UCP manager.
+#   Only required if you are using UCP 2.0 and above
+#
 # [*ucp_id*]
 #   The ID for the UCP. Used when deleting UCP with ensure => absent.
 #
 # [*fingerprint*]
 #   The certificate fingerprint for the UCP controller.
 #   Required for nodes.
+#
+# [*token*]
+#  This is the authtentication token used for UCP 2.0 and above
+#  Required only if you are using UCP version 2.0 or higher
 #
 # [*replica*]
 #   Whether or not this is a replica of the controller. Defaults to false.
@@ -107,18 +115,22 @@ class docker_ucp(
   $preserve_certs_on_delete = false,
   $preserve_images_on_delete = false,
   $ucp_url = undef,
+  $ucp_manager = undef,
   $ucp_id = undef,
   $fingerprint = undef,
+  $token = undef,
+  $listen_address = undef,
+  $advertise_address = undef,
   $replica = false,
   $username = 'admin',
   $password = 'orca',
   $license_file = undef,
   $local_client = false,
 ) {
-  validate_re($::osfamily, '^(Debian|RedHat)$', "${::operatingsystem} not supported. This module only works on Debian and Red Hat based systems.")
+  validate_re($::osfamily, '^(Debian|RedHat)$', "${::operatingsystem} not supported. This module only works on Debian and Red Hat based systems.") # lint:ignore:140chars
 
   validate_re($ensure, '^(present|absent)$')
-  validate_bool($tracking, $usage, $preserve_certs, $preserve_certs_on_delete, $preserve_images_on_delete, $controller, $external_ca, $replica)
+  validate_bool($tracking, $usage, $preserve_certs, $preserve_certs_on_delete, $preserve_images_on_delete, $controller, $external_ca, $replica) # lint:ignore:140chars
   validate_absolute_path($docker_socket_path)
   validate_string($host_address, $version, $ucp_url, $ucp_id, $fingerprint, $username, $password)
 
@@ -180,6 +192,8 @@ class docker_ucp(
   } else {
     if $controller {
       $install_flags = ucp_install_flags({
+        admin_username     => $username,
+        admin_password     => $password,
         host_address       => $host_address,
         tracking           => $tracking,
         usage              => $usage,
@@ -197,7 +211,7 @@ class docker_ucp(
       })
       if $license_file {
         exec { 'Install Docker Universal Control Plane':
-          command => "docker run --rm -v ${docker_socket_path}:/var/run/docker.sock -v ${license_file}:/docker_subscription.lic --name ucp docker/ucp install ${install_flags}",
+          command => "docker run --rm -v ${docker_socket_path}:/var/run/docker.sock -v ${license_file}:/docker_subscription.lic --name ucp docker/ucp install ${install_flags}", # lint:ignore:140chars
           unless  => $install_unless,
         }
       } else {
@@ -221,9 +235,19 @@ class docker_ucp(
         san                => any2array($subject_alternative_names),
         extra_parameters   => any2array($extra_parameters),
       })
-      exec { 'Join Docker Universal Control Plane':
-        command => "docker run --rm -v ${docker_socket_path}:/var/run/docker.sock -e 'UCP_ADMIN_USER=${username}' -e 'UCP_ADMIN_PASSWORD=${password}' --name ucp docker/ucp join ${join_flags}",
-        unless  => $join_unless,
+
+      if $version =~ /^2.*/ {
+        exec { 'Join Docker Universal Control Plane v2':
+          command => "docker swarm join --listen-addr ${listen_address} --advertise-addr ${advertise_address}:2377  --token ${token} ${ucp_manager}:2377", # lint:ignore:140chars
+          unless  => $join_unless,
+          }
+      }
+
+      else {
+        exec { 'Join Docker Universal Control Plane v1':
+          command => "docker run --rm -v ${docker_socket_path}:/var/run/docker.sock -e 'UCP_ADMIN_USER=${username}' -e 'UCP_ADMIN_PASSWORD=${password}' --name ucp docker/ucp join ${join_flags}", # lint:ignore:140chars
+          unless  => $join_unless,
+        }
       }
     }
   }
